scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
youtube-transcript-api==0.6.1